requirements.txt
fastapi
uvicorn
pystac-client
rasterio
numpy
sqlalchemy
geoalchemy2
asyncpg
alembic
python-dotenv
pydantic
typing
dotenv
matplotlib
cors
requests
Pillow
torch
segmentation-models-pytorch
pyproj
